refactor: ProductService, ProductRestController 새로운 Product 도메인에 맞게 수정
